rearrange components for feedback pages, clean up duplications
