+++ src/js/components/App.js
@@ -1,11 +1,15 @@
 /*global SEGMENT_KEY:true*/
 
-import React, { PropTypes } from 'react'
-import Header from './common/Header'
-import Footer from './common/Footer'
+import React, { PropTypes } from 'react';
+import Header from './common/Header';
+import Footer from './common/Footer';
+import toastr from 'toastr';
 
 class App extends React.Component {
   render () {
+    toastr.options.positionClass = "toast-nav-top-right";
+    toastr.options.preventDuplicates = true;
+
     return (
       <div>
         <Header />
