+++ src/js/components/feedback/success/Application.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import {Link} from 'react-router';
+import logo from '../../../../img/logo-colored.png';
+
+const ApplicationSuccess = () => {
+  return (
+    <div className="page-title">
+      <h2>We've received your application!</h2>
+      <p>Thank you for reaching out to partner with us. We will be reaching out shortly by sending you an email or a text message.</p>
+    </div>
+  )
+}
+
+export default ApplicationSuccess;
