+++ src/js/components/feedback/success/ContactUs.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import logo from '../../../../img/logo-colored.png';
+import {Link} from 'react-router';
+
+const ContactUsSuccess = () => {
+  return (
+    <div className="page-title">
+      <h2>Thank you for reaching out!</h2>
+      <p>We have received your message successfully and will get back to you as soon as we can.</p>
+    </div>
+  )
+}
+
+export default ContactUsSuccess;
