+++ src/js/components/partner_leads/Index.js
@@ -1,7 +1,11 @@
 import React, {PropTypes} from 'react';
+import toastr from 'toastr';
 
 class PartnerLeads extends React.Component {
   render() {
+    toastr.options.positionClass = "toast-top-right";
+    toastr.options.preventDuplicates = true;
+
     return (
       <div className="partner-leads">
         {this.props.children}
