+++ src/js/components/partner_leads/eligibility_check/Index.js
@@ -3,6 +3,7 @@ import logo from '../../../../img/logo-colored.png';
 import {Link, browserHistory} from 'react-router';
 import SingleQuestion from './SingleQuestion';
 import criteriaList from './criteriaList';
+import DocumentTitle from '../../common/DocumentTitle';
 
 class EligibilityCheck extends React.Component {
   constructor() {
@@ -52,7 +53,7 @@ class EligibilityCheck extends React.Component {
       return false;
     });
     if (percentageCompletion != 100) return;
-    browserHistory.push(eligible ? '/partner-leads/apply' : '/partner-leads/ineligible');
+    browserHistory.push(eligible ? '/partner-leads/apply' : '/feedback/error/ineligible');
   };
 
   render() {
@@ -92,4 +93,4 @@ class EligibilityCheck extends React.Component {
   }
 }
 
-export default EligibilityCheck;
+export default DocumentTitle('Eligibility')(EligibilityCheck);
