+++ src/js/routes.js
@@ -4,14 +4,14 @@ import App from './components/App';
 import HomePage from './components/home/HomePage';
 import CurriculumPage from './components/curriculum/Index';
 import ApplicationPage from './components/partner_leads/application/Index';
-import IneligiblePage from './components/partner_leads/ineligible/Index';
-import AppreciationPage from './components/partner_leads/appreciation/Index';
+import EligibilityErrorPage from './components/feedback/error/Eligibility';
+import ApplicationSuccessPage from './components/feedback/success/Application';
 import EligibilityCheck from './components/partner_leads/eligibility_check/Index';
-import PartnerLeadsPage from './components/partner_leads/Index';
 import * as flagChecks from './utils/featureFlagChecks';
-import ContactUs from './components/contact_us/Index';
-import ContactUsSuccessPage from './components/contact_us/ContactSuccess';
-import ContactUsPage from './components/contact_us/ContactUsPage';
+import PartnerLeads from './components/partner_leads/Index';
+import ContactUsSuccessPage from './components/feedback/success/ContactUs';
+import Feedback from './components/feedback/Index';
+import ContactUsPage from './components/contact_us/Index';
 
 const redirectIfFlagIsDisabled = (flag, nextState, replaceState) => {
   if (flag) return;
@@ -28,20 +28,18 @@ export default (
       <Route path="contact-us"
              component={ContactUsPage}
              onEnter={redirectIfFlagIsDisabled.bind(null, flagChecks.contactUsIsEnabled())} />
-      <Route path="feedback" 
-             component={ContactUsPage}
-             onEnter={redirectIfFlagIsDisabled.bind(null, flagChecks.contactUsIsEnabled())} />
     </Route>
 
-    <Route path="/contact-us" component={ContactUs} >
-      <Route path="success" component={ContactUsSuccessPage} />
+    // TODO: Create template for feedback and pass in custom messages
+    <Route path="/feedback" component={Feedback} >
+      <Route path="success/application" component={ApplicationSuccessPage} />
+      <Route path="success/contact-us" component={ContactUsSuccessPage} />
+      <Route path="error/ineligible" component={EligibilityErrorPage} />
     </Route>
 
-    <Route path="/partner-leads" component={PartnerLeadsPage} >
+    <Route path="/partner-leads" component={PartnerLeads} >
       <Route path="apply" component={ApplicationPage} />
-      <Route path="thank-you" component={AppreciationPage} />
       <Route path="check-eligibility" component={EligibilityCheck} />
-      <Route path="ineligible" component={IneligiblePage} />
     </Route>
   </Route>
 )
