+++ src/scss/4-pages/contact_us.scss
@@ -19,6 +19,7 @@
     }
 }
 
+.back-to-home,
 .contact-form-submit {
     margin-top: 60px;
     margin-bottom: 10px;
