+++ src/scss/4-pages/shared.scss
@@ -23,7 +23,7 @@
     width: 100%;
 
     img {
-      height: 100%;
+      height: 90%;
 
       @include mobile {
         height: 70%;
@@ -58,6 +58,7 @@
     }
     p {
       font-size: 22px;
+      color: #555555;
       margin-top: 50px;
     }
 
@@ -66,19 +67,24 @@
       font-size: 0.8rem;
     }
   }
+}
 
-  .wrapper {
-    p {
-      font-size: 28px;
-      margin-top: 60px;
-      margin-bottom: 80px;
-      text-align: center;
-      color: #999999;
+.toast-nav-top-right {
+  top: 65px;
+  right: 12px;
+}
 
-      @include mobile {
-        font-size: 28px;
-        line-height: 1.5;
-      }
-    }
+#toast-container > div {
+  border-radius: 0;
+  -moz-box-shadow: 0 0 12px #383434;
+  -webkit-box-shadow: 0 0 12px #383434;
+  box-shadow: 0 0 12px #383434;
+}
+
+#toast-container > .toast-error {
+  background-image: none !important;
+
+  .toast-message {
+    margin-left: -30px;
   }
 }
